Upgrade to Flask 0.12.2 + switch to Flask Blueprints
The Flask Module system we were using in LNT was depricated and removed
from Flask long ago. This changes us from Module to Blueprints. Biggest
change is that all url_for calls need to specify the blueprint name,
or add a "." for the local blueprint.  I updated all uses of url_for
to correctly account for this.